VoronoiFVMDiffEq.jl

Glue package between VoronoiFVM.jl and DifferentialEquations.jl

The package extends the constructors for ODEFunction and ODEProblem by methods taking in a VoronoiFVM.System:

problem=ODEProblem(sys::VoronoiFVM.System,inival,tspan,callback=DifferentialEquations.CallbackSet())
odesolution=DifferentialEquations.solve(problem)
voronoifvmsolution=reshape(odesolution, sys::VoronoiFVM.System)

It re-exports VoronoiFVM, so that it is sufficient to use this package instead of VoronoiFVM.

The package requires a recent Julia version (currently 1.8) due to significant advances in package precompilation.

This package is in pre-release state. Versions 0.0.x are registered in https://github.com/j-fu/PackageNursery. If you trust this registry, you can issue once

pkg> registry add https://github.com/j-fu/PackageNursery

and add the package via

pkg> add VoronoiFVMDiffEq

Choice of ODE/DAE solver

As this package interfaces to the PDE solver package VoronoiFVM.jl, a general advise is to choose methods able to handle stiff problems. Moreover, often, discretized PDE systems (e.g. containing elliptic equations) are differential agebraic equation (DAE) systems which should be solved by DAE solvers.

Methods

SciMLBase.ODEProblem — Type

ODEProblem(system,inival,tspan,callback=DifferentialEquations.CallbackSet())

Create an ODEProblem which can be handeled by ODE solvers from DifferentialEquations.jl.

Parameters:

system: A VoronoiFVM.System

inival: Initial value. Consider to pass a stationary solution at tspan[1].

tspan: Time interval

callback : (optional) callback for ODE solver

The method returns an ODEProblem which can be solved by DifferentialEquations.solve().

SciMLBase.ODEFunction — Type

 ODEFunction(system,inival=unknowns(system,inival=0),t0=0)

Create an ODEPFunction which can be handeled by ODE solvers from DifferentialEquations.jl.

Parameters:

system: A VoronoiFVM.System

jacval (optional): Initial value. Default is a zero vector. Consider to pass a stationary solution at time tjac.

tjac (optional): tjac, Default: 0

The jacval and tjac are passed for a first evaluation of the Jacobian, allowing to detect the sparsity pattern which is passed to the DifferentialEquations.jl solver.

Base.reshape — Function

reshape(ode_solution, system)

Create a TransientSolution from the output of the ode solver.